[REFACTOR] move crypto.cpp to network/legacy
reason: hashpassword and checkpassword needed by both client and server

additionally: remove auth crypto code
