+++ inexor/shared/iengine.hpp
@@ -464,15 +464,6 @@ extern void neterr(const char *s, bool disc = true);
 extern void gets2c();
 extern void notifywelcome();
 
-// crypto
-extern void genprivkey(const char *seed, vector<char> &privstr, vector<char> &pubstr);
-extern bool hashstring(const char *str, char *result, int maxlen);
-extern void answerchallenge(const char *privstr, const char *challenge, vector<char> &answerstr);
-extern void *parsepubkey(const char *pubstr);
-extern void freepubkey(void *pubkey);
-extern void *genchallenge(void *pubkey, const void *seed, int seedlen, vector<char> &challengestr);
-extern void freechallenge(void *answer);
-extern bool checkchallenge(const char *answerstr, void *correct);
 
 // 3dgui
 struct Texture;
